+++ src/app/components/admin/products/product-detail/product-detail.component.ts
@@ -37,12 +37,17 @@ export class ProductDetailComponent implements OnInit {
   }
 
   async onGoToDelete(): Promise<void> {
-    try {
-      await this.productSvc.onDeleteProduct(this.product.id);
-      alert('delete');
-      this.onGoBackToList();
-    } catch (err) {
-      console.log(err);
+    const confirmacion = confirm('are you sure you want to delete it?');
+    if (confirmacion) {
+      try {
+        await this.productSvc.onDeleteProduct(this.product.id);
+        alert('delete');
+        this.onGoBackToList();
+      } catch (err) {
+        console.log(err);
+      }
+    } else {
+      return;
     }
   }
 }
